test: update test structure to conform with azure and gcp  (PSKD-1475)
