+++ test/defaultplan/networking_test.go
@@ -27,6 +27,33 @@ func TestPlanNetworking(t *testing.T) {
 			ResourceMapName:   "module.vpc.aws_subnet.public[0]",
 			AttributeJsonPath: "{$.availability_zone}",
 		},
+		// Use Existing
+		/*
+			"vnetSubnetTest": {
+				Expected:          "",
+				ResourceMapName:   "module.vnet.azurerm_virtual_network.vnet[0]",
+				AttributeJsonPath: "{$.subnet[0].name}",
+			},
+			"clusterEgressTypeTest": {
+				Expected:          "loadBalancer",
+				ResourceMapName:   "module.aks.azurerm_kubernetes_cluster.aks",
+				AttributeJsonPath: "{$.network_profile[0].outbound_type}",
+			},
+			"networkPluginTest": {
+				Expected:          "kubenet",
+				ResourceMapName:   "module.aks.azurerm_kubernetes_cluster.aks",
+				AttributeJsonPath: "{$.network_profile[0].network_plugin}",
+			},
+			"aksNetworkPolicyTest": {
+				Expected:          "",
+				ResourceMapName:   "module.aks.azurerm_kubernetes_cluster.aks",
+				AttributeJsonPath: "{$.expressions.aks_network_policy.reference[0]}",
+			},
+			"aksNetworkPluginModeTest": {
+				Expected:          "",
+				ResourceMapName:   "module.aks.azurerm_kubernetes_cluster.aks",
+				AttributeJsonPath: "{$.expressions.aks_network_plugin_mode.reference[0]}",
+			},*/
 	}
 
 	helpers.RunTests(t, tests, helpers.GetDefaultPlan(t))
